new components, navigation, clean script
