+++ app/views/welcome/index.html
@@ -1,3 +1,4 @@
+<base href="/">
 <script src="/systemjs.config.js"></script>
 <script>
 System.import('app').catch(function(err){ console.error(err); });
