+++ clean.sh
@@ -1,0 +1,4 @@
+mv public/systemjs.config.js public/systemjs.config.js.bak
+rm public/*.js.map
+rm public/*.js
+mv public/systemjs.config.js.bak public/systemjs.config.js
