+++ ng-app/app.component.ts
@@ -2,11 +2,7 @@ import { Component } from '@angular/core';
 
 @Component({
   selector: 'my-app',
-  template: `
-    <h1>{{title}}</h1>
-    <my-workouts></my-workouts>
-  `,
-  styleUrls: ['']
+  templateUrl: './app.component.html',
 })
 export class AppComponent {
   title = 'Workout Builder';
