+++ ng-app/app.module.ts
@@ -3,18 +3,22 @@ import { BrowserModule } from '@angular/platform-browser';
 import { FormsModule }   from '@angular/forms';
 
 import { AppComponent }        from './app.component';
-import { WorkoutComponent } from './workout.component';
+import { WorkoutsComponent } from './workouts.component';
+import { HomeComponent } from './home.component';
 import { WorkoutDetailComponent } from './workout-detail.component';
+import { AppRoutingModule } from './app-routing.module'
 
 @NgModule({
   imports: [
     BrowserModule,
-    FormsModule
+    FormsModule,
+    AppRoutingModule
   ],
   declarations: [
     AppComponent,
-    WorkoutComponent,
-    WorkoutDetailComponent
+    WorkoutsComponent,
+    WorkoutDetailComponent,
+    HomeComponent
   ],
   bootstrap: [ AppComponent ]
 })
