+++ ng-app/home.component.ts
@@ -1,0 +1,11 @@
+import { Component, Input } from '@angular/core';
+
+
+@Component({
+  selector: 'home',
+  templateUrl: './home.component.html'
+})
+
+export class HomeComponent {
+
+}
