+++ ng-app/workout.ts
@@ -1,4 +1,11 @@
 export class Workout {
   id: number;
   name: string;
+  // exerciseType: string;
+  // description: string;
+  // sets: number;
+  // reps: number;
+  // rest: number;
+  // duration: number;
+  // distance: string;
 }
