+++ public/home.component.html
@@ -1,0 +1,4 @@
+<div id="home-component">
+  <h4>Create your Workout, search below:</h4>
+  <input #searchBox id="search-box" />
+</div>
