+++ public/workout-detail.component.html
@@ -1,0 +1,8 @@
+<div *ngIf="workout">
+  <h2>{{workout.name}} details!</h2>
+  <div><label>id: </label>{{workout.id}}</div>
+  <div>
+    <label>name: </label>
+    <input [(ngModel)]="workout.name" placeholder="name"/>
+  </div>
+</div>
