+++ public/workouts.component.css
@@ -18,14 +18,14 @@
   height: 1.6em;
   border-radius: 4px;
 }
+.workouts li.selected:hover {
+  background-color: #BBD8DC !important;
+  color: white;
+}
 .workouts li:hover {
   color: #607D8B;
   background-color: #DDD;
   left: .1em;
-}
-.workouts li.selected:hover {
-  background-color: #BBD8DC !important;
-  color: white;
 }
 .workouts .text {
   position: relative;
